+++ CHANGELOG.md
@@ -8,4 +8,5 @@
   client detail.
 - **[Management API - Users]** Implemented endpoints list users and get user
   detail.
-- **[Authentication API]** Implemented client credentials flow.
+- **[Authentication API]** Implemented client credentials flow and resource
+  owner password flow.